% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slave.R
\name{finish_task}
\alias{finish_task}
\title{Send result back to QWOP task master}
\usage{
finish_task(master_root, task_id, score)
}
\arguments{
\item{master_root}{Root url of the QWOP task master}

\item{task_id}{ID of the completed task}

\item{score}{Score on the task}
}
\value{
success (bool)
}
\description{
\code{finish_task} Sends HTTP POST back to task master, with the score for a completed task
}

