% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesseract.R
\name{get_score}
\alias{get_score}
\title{Parse a score with Tesseract OCR}
\usage{
get_score(png_file, screenshot_dir = "screenshots", output_dir = "outputs",
  alive_factor = 1.2)
}
\arguments{
\item{png_file}{Path to .png image of the end of a QWOP run}

\item{screenshot_dir}{location of screenshot .pngs}

\item{output_dir}{location to save OCR outputs}

\item{alive_factor}{Bonus multiplier for QWOP runs that are still alive at the end}
}
\value{
Score from a QWOP run
}
\description{
\code{get_score} Uses Tesseract OCR to find the score from a snapshot.
 Assumes that you have tesseract installed and in the PATH
}

