% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slave.R
\name{get_task}
\alias{get_task}
\title{Get a task from the QWOP task master}
\usage{
get_task(master_root)
}
\arguments{
\item{master_root}{Root url of the QWOP task master
See source on GitHub for master instructions}
}
\value{
list(id, sequence)
}
\description{
\code{get_task} Sends HTTP GET to the task manager, gets and parses the task's JSON
}

