% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{run_sequence}
\alias{run_sequence}
\title{Run a sequence of moves}
\usage{
run_sequence(driver, sequence, screenshot_dir = "screenshots")
}
\arguments{
\item{driver}{QWOP driver created by \code{initialize_driver}}

\item{sequence}{(nx4) matrix specifying a sequence of Q, W, O, and P actions}

\item{screenshot_dir}{Location to save screenshots to}
}
\value{
Path to the snapshot (.png) of the game after running the sequence
}
\description{
\code{run_sequence} Runs a sequence of moves
}

